+++ src/main/kotlin/at/asitplus/gradle/IosFrameworkSetup.kt
@@ -3,18 +3,27 @@ package at.asitplus.gradle
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.getByType
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.plugin.mpp.Framework
+import org.jetbrains.kotlin.gradle.plugin.mpp.BitcodeEmbeddingMode
 import org.jetbrains.kotlin.gradle.plugin.mpp.apple.XCFrameworkConfig
 
 
-fun Project.exportIosFramework(name: String, vararg additionalExport: Any){
-    extensions.getByType<KotlinMultiplatformExtension>().apply{
+fun Project.exportIosFramework(
+    name: String,
+    vararg additionalExports: Any
+) = exportIosFramework(name, bitcodeEmbeddingMode = BitcodeEmbeddingMode.BITCODE, additionalExports = additionalExports)
+
+fun Project.exportIosFramework(
+    name: String,
+    bitcodeEmbeddingMode: BitcodeEmbeddingMode,
+    vararg additionalExports: Any
+) {
+    extensions.getByType<KotlinMultiplatformExtension>().apply {
         XCFrameworkConfig(project, name).also { xcf ->
             ios {
                 binaries.framework {
                     baseName = name
-                    embedBitcode("bitcode")
-                    additionalExport.forEach {
+                    embedBitcode(bitcodeEmbeddingMode)
+                    additionalExports.forEach {
                         export(it)
                     }
                     xcf.add(this)
@@ -23,8 +32,8 @@ fun Project.exportIosFramework(name: String, vararg additionalExport: Any){
             iosSimulatorArm64 {
                 binaries.framework {
                     baseName = name
-                    embedBitcode("bitcode")
-                    additionalExport.forEach {
+                    embedBitcode(bitcodeEmbeddingMode)
+                    additionalExports.forEach {
                         export(it)
                     }
                     xcf.add(this)